+++ source/extensions/omni.isaac.dofbot/omni/isaac/dofbot/__init__.py
@@ -7,4 +7,4 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 from omni.isaac.dofbot.dofbot import DofBot
-from omni.isaac.dofbot.inverse_kinematics_solver import InverseKinematicsSolver
+from omni.isaac.dofbot.kinematics_solver import KinematicsSolver