+++ source/extensions/omni.isaac.motion_generation/omni/isaac/motion_generation/__init__.py
@@ -12,6 +12,9 @@ from omni.isaac.motion_generation.articulation_subset import ArticulationSubset
 from omni.isaac.motion_generation.world_interface import WorldInterface
 from omni.isaac.motion_generation.motion_policy_interface import MotionPolicy
 from omni.isaac.motion_generation.lula.motion_policies import RmpFlow, RmpFlowSmoothed
+from omni.isaac.motion_generation.lula.kinematics import LulaKinematicsSolver
+from omni.isaac.motion_generation.articulation_kinematics_solver import ArticulationKinematicsSolver
+from omni.isaac.motion_generation.kinematics_interface import KinematicsSolver
 from omni.isaac.motion_generation.motion_policy_controller import MotionPolicyController
 from omni.isaac.motion_generation.pick_place_controller import PickPlaceController
 from omni.isaac.motion_generation.stacking_controller import StackingController