+++ source/extensions/omni.isaac.motion_generation/omni/isaac/motion_generation/lula/__init__.py
@@ -1,1 +1,2 @@
 from .motion_policies import RmpFlow
+from .kinematics import LulaKinematicsSolver