+++ source/extensions/omni.isaac.motion_generation/omni/isaac/motion_generation/tests/__init__.py
@@ -8,3 +8,4 @@
 #
 
 from .test_motion_policy import *
+from .test_kinematics import *