+++ source/standalone_examples/api/omni.isaac.dofbot/follow_target_with_ik.py
@@ -10,24 +10,23 @@ from omni.isaac.kit import SimulationApp
 
 simulation_app = SimulationApp({"headless": False})
 
-from omni.isaac.dofbot.tasks import PickPlace
-from omni.isaac.dofbot.controllers import PickPlaceController
+from omni.isaac.dofbot.tasks import FollowTarget
+from omni.isaac.dofbot.controllers import RMPFlowController
 from omni.isaac.core import World
+from omni.isaac.dofbot import KinematicsSolver
+import carb
 import numpy as np
 
 my_world = World(stage_units_in_meters=0.01)
-my_task = PickPlace()
+my_task = FollowTarget(name="follow_target_task")
 my_world.add_task(my_task)
 my_world.reset()
-task_params = my_task.get_params()
+task_params = my_world.get_task("follow_target_task").get_params()
 dofbot_name = task_params["robot_name"]["value"]
+target_name = task_params["target_name"]["value"]
 my_dofbot = my_world.scene.get_object(dofbot_name)
-my_controller = PickPlaceController(
-    name="pick_place_controller", gripper_dof_indices=my_dofbot.gripper.dof_indices, robot_articulation=my_dofbot
-)
+my_controller = KinematicsSolver(my_dofbot)
 articulation_controller = my_dofbot.get_articulation_controller()
-
-i = 0
 while simulation_app.is_running():
     my_world.step(render=True)
     if my_world.is_playing():
@@ -35,14 +34,14 @@ while simulation_app.is_running():
             my_world.reset()
             my_controller.reset()
         observations = my_world.get_observations()
-        actions = my_controller.forward(
-            picking_position=observations[task_params["cube_name"]["value"]]["position"],
-            placing_position=observations[task_params["cube_name"]["value"]]["target_position"],
-            current_joint_positions=observations[dofbot_name]["joint_positions"],
-            end_effector_offset=np.array([0, -6, 0]),
-        )
-        if my_controller.is_done():
-            print("done picking and placing")
-        articulation_controller.apply_action(actions)
+
+        # IK does not work well on dofbot with orientation targets
+        actions, succ = my_controller.compute_inverse_kinematics(target_position=observations[target_name]["position"])
+        # actions,succ = my_controller.compute_inverse_kinematics(target_position=observations[target_name]["position"],
+        #     target_orientation=observations[target_name]["orientation"], orientation_tolerance = np.pi/2)
+        if succ:
+            articulation_controller.apply_action(actions)
+        else:
+            carb.log_warn("IK did not converge to a solution.  No action is being taken.")
 
 simulation_app.close()