+++ source/extensions/omni.isaac.dofbot/docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 CHANGELOG
 **********
 
+[0.2.0] - 2022-05-02
+========================
+
+Changed
+------
+- Changed InverseKinematicsSolver class to KinematicsSolver class, using the new LulaKinematicsSolver class in motion_generation
+
 [0.1.4] - 2022-04-21
 ========================